Make list operation utility component
