+++ ui/v2.5/src/components/List/ListOperationButtons.tsx
@@ -6,7 +6,10 @@ import { IconDefinition } from "@fortawesome/fontawesome-svg-core";
 import { Icon } from "../Shared/Icon";
 import {
   faEllipsisH,
+  faPencil,
   faPencilAlt,
+  faPlay,
+  faPlus,
   faTrash,
 } from "@fortawesome/free-solid-svg-icons";
 import cx from "classnames";
@@ -262,5 +265,110 @@ export const ListOperationButtons: React.FC<IListOperationButtonsProps> = ({
         {moreDropdown}
       </ButtonGroup>
     </>
+  );
+};
+
+interface IListOperations {
+  text: string;
+  onClick: () => void;
+  isDisplayed?: () => boolean;
+  className?: string;
+}
+
+export const ListOperations: React.FC<{
+  items: number;
+  hasSelection?: boolean;
+  operations?: IListOperations[];
+  onEdit?: () => void;
+  onDelete?: () => void;
+  onPlay?: () => void;
+  onCreateNew?: () => void;
+  entityType?: string;
+  operationsClassName?: string;
+  operationsMenuClassName?: string;
+}> = ({
+  items,
+  hasSelection = false,
+  operations = [],
+  onEdit,
+  onDelete,
+  onPlay,
+  onCreateNew,
+  entityType,
+  operationsClassName,
+  operationsMenuClassName,
+}) => {
+  const intl = useIntl();
+
+  return (
+    <div className="list-operations">
+      <ButtonGroup>
+        {!!items && onPlay && (
+          <Button
+            className="play-button"
+            variant="secondary"
+            onClick={() => onPlay()}
+            title={intl.formatMessage({ id: "actions.play" })}
+          >
+            <Icon icon={faPlay} />
+          </Button>
+        )}
+        {!hasSelection && onCreateNew && (
+          <Button
+            className="create-new-button"
+            variant="secondary"
+            onClick={() => onCreateNew()}
+            title={intl.formatMessage(
+              { id: "actions.create_entity" },
+              { entityType }
+            )}
+          >
+            <Icon icon={faPlus} />
+          </Button>
+        )}
+
+        {hasSelection && (onEdit || onDelete) && (
+          <>
+            {onEdit && (
+              <Button variant="secondary" onClick={() => onEdit()}>
+                <Icon icon={faPencil} />
+              </Button>
+            )}
+            {onDelete && (
+              <Button
+                variant="danger"
+                className="btn-danger-minimal"
+                onClick={() => onDelete()}
+              >
+                <Icon icon={faTrash} />
+              </Button>
+            )}
+          </>
+        )}
+
+        {operations.length > 0 && (
+          <OperationDropdown
+            className={operationsClassName}
+            menuClassName={operationsMenuClassName}
+            menuPortalTarget={document.body}
+          >
+            {operations.map((o) => {
+              if (o.isDisplayed && !o.isDisplayed()) {
+                return null;
+              }
+
+              return (
+                <OperationDropdownItem
+                  key={o.text}
+                  onClick={o.onClick}
+                  text={o.text}
+                  className={o.className}
+                />
+              );
+            })}
+          </OperationDropdown>
+        )}
+      </ButtonGroup>
+    </div>
   );
 };
